services/profile/pyproject.toml
[tool.poetry]
name = "budget-tracker-profile"
version = "0.1.0"
description = "Profile microservice for budget-tracker."
authors = ["simon.islam <simon.islam@brilliant.com.bd>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
flask-restx = "^0.5.1"
gunicorn = "^20.1.0"
requests = "^2.27.1"
Flask-Migrate = "^3.1.0"
PyJWT = "^2.3.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.2"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
